initial commit for the action (#1)
<!-- markdownlint-disable-next-line first-line-heading -->
## Description
The initial commit - copying the format from
[actions/typescript-action](https://github.com/actions/typescript-action)
to implement the framework plus updating main.ts to enable the sending
of a notification to MS Teams.

## Context

This is to enable teams to easily send notifications from their build
pipelines to Microsoft Teams.

## Type of changes

- [ ] Refactoring (non-breaking change)
- [X] New feature (non-breaking change which adds functionality)
- [X] Breaking change (fix or feature that would change existing
functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)

## Checklist

- [X] I am familiar with the [contributing
guidelines](../docs/CONTRIBUTING.md)
- [X] I have followed the code style of the project
- [^] I have added tests to cover my changes
- [X] I have updated the documentation accordingly
- [ ] This PR is a result of pair or mob programming

---

## Sensitive Information Declaration

To ensure the utmost confidentiality and protect your and others
privacy, we kindly ask you to NOT including [PII (Personal Identifiable
Information) / PID (Personal Identifiable
Data)](https://digital.nhs.uk/data-and-information/keeping-data-safe-and-benefitting-the-public)
or any other sensitive data in this PR (Pull Request) and the codebase
changes. We will remove any PR that do contain any sensitive
information. We really appreciate your cooperation in this matter.

- [X] I confirm that neither PII/PID nor sensitive data are included in
this PR and the codebase changes.